pyproject.toml
[project]
name = "preferences"
version = "0.1.0"
description = "A library for measuring AI model preferences"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "instructor>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.0.0",
    "autograd>=1.6.0",
    "rich>=13.0.0",
    "scipy>=1.11.0",
    "numpy>=1.26.4",
    "scikit-learn>=1.3.0",
    "nnsight>=0.2.11",
    "transformer-lens>=2.17.0",
    "transformers>=4.57.0,<5.0.0",
    "huggingface-hub>=0.36.0,<1.0.0",
    "tenacity>=8.0.0",
    "trueskill>=0.4.5",
    "vllm>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.0",
]

viz = [
    "adjustText>=1.0.0",
    "ipywidgets>=8.0.0",
    "matplotlib>=3.5.0",
    "mplcursors>=0.5.0",
    "notebook>=7.0.0",
    "pandas>=2.0.0",
    "plotly>=5.0.0",
    "seaborn>=0.13.0",
    "IPython>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-m 'not api and not gpu'"
markers = [
    "api: tests that make real API calls (requires API key)",
    "gpu: tests that require a GPU or heavy model loading",
    "slow: tests that take a long time to run",
    "models: tests for LLM abstraction and client implementations",
    "tasks: tests for task loading and dataset handling",
    "prompts: tests for prompt templates, builders, and generation",
    "measurement: tests for preference measurers and response parsing",
    "cache: tests for measurement storage and caching",
    "runners: tests for experiment runners and orchestration",
    "probes: tests for activation extraction and linear probes",
    "thurstonian: tests for Thurstonian model fitting and active learning",
    "trueskill: tests for TrueSkill model fitting",
    "ranking: tests for ranking-based preference measurement",
    "analysis: tests for post-hoc analysis modules",
    "steering: tests for activation steering experiments",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["FA"]  # flake8-future-annotations
